requirements.txt
streamlit>=1.32
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
matplotlib>=3.7
reportlab>=4.0
